[LibOS] Split off x86_64 specifics from shim_tcb.h
Split off x86_64 specifics from shim_tcb.h and move them into
include/arch/x86_64/shim_tcb-arch.h.
