+++ LibOS/shim/include/arch/x86_64/shim_tcb-arch.h
@@ -1,0 +1,27 @@
+#ifndef _SHIM_TCB_ARCH_H_
+#define _SHIM_TCB_ARCH_H_
+
+#include <stdint.h>
+
+struct shim_regs {
+    uint64_t    orig_rax;
+    uint64_t    rsp;
+    uint64_t    r15;
+    uint64_t    r14;
+    uint64_t    r13;
+    uint64_t    r12;
+    uint64_t    r11;
+    uint64_t    r10;
+    uint64_t    r9;
+    uint64_t    r8;
+    uint64_t    rcx;
+    uint64_t    rdx;
+    uint64_t    rsi;
+    uint64_t    rdi;
+    uint64_t    rbx;
+    uint64_t    rbp;
+    uint64_t    rflags;
+    uint64_t    rip;
+};
+
+#endif /* _SHIM_TCB_ARCH_H_ */
