+++ LibOS/shim/include/shim_tcb.h
@@ -5,28 +5,9 @@
 #include <atomic.h>
 #include <pal.h>
 
-#define SHIM_TCB_CANARY 0xdeadbeef
+#include "shim_tcb-arch.h"
 
-struct shim_regs {
-    uint64_t    orig_rax;
-    uint64_t    rsp;
-    uint64_t    r15;
-    uint64_t    r14;
-    uint64_t    r13;
-    uint64_t    r12;
-    uint64_t    r11;
-    uint64_t    r10;
-    uint64_t    r9;
-    uint64_t    r8;
-    uint64_t    rcx;
-    uint64_t    rdx;
-    uint64_t    rsi;
-    uint64_t    rdi;
-    uint64_t    rbx;
-    uint64_t    rbp;
-    uint64_t    rflags;
-    uint64_t    rip;
-};
+#define SHIM_TCB_CANARY 0xdeadbeef
 
 struct shim_context {
     struct shim_regs *      regs;
